Move frontend files into their own directory
All the frontend files have now been moved to their own directory.
