+++ arm_vm.c
@@ -20,7 +20,7 @@
 
 #include "arm_disass.h"
 #include "arm_vm.h"
-#include "utils.h"
+#include "common/utils.h"
 
 subtilis_arm_vm_t *subtilis_arm_vm_new(uint32_t *code, size_t code_size,
 				       size_t mem_size, subtilis_error_t *err)
