+++ check-format.sh
@@ -5,7 +5,7 @@ if [ -f /usr/bin/clang-format-3.9 ]; then
     CLANG_FORMAT=/usr/bin/clang-format-3.9
 fi
 $CLANG_FORMAT --version
-for fn in *.[ch] common/*.[ch]; do
+for fn in *.[ch] common/*.[ch] frontend/*.[ch] test_cases/*.[ch]; do
     echo ${fn}
     $CLANG_FORMAT -style=file ${fn} | diff ${fn}  -
 done
