+++ frontend/globals.c
