+++ frontend/globals.h
