+++ frontend/stream.c
