+++ frontend/symbol_table.h
