+++ unit_tests.c
@@ -21,10 +21,10 @@
 #include "arm_test.h"
 #include "common/bitset_test.h"
 #include "fpa_test.h"
-#include "ir_test.h"
-#include "lexer_test.h"
-#include "parser_test.h"
-#include "symbol_table_test.h"
+#include "frontend/ir_test.h"
+#include "frontend/lexer_test.h"
+#include "frontend/parser_test.h"
+#include "frontend/symbol_table_test.h"
 
 int main(int argc, char *argv[])
 {
